Changed db phene cuttoffs from int to doubles and fixed workflow issues

+++ src/main/java/cfe/action/ClinicalAndTestingCohortsAction.java
@@ -88,8 +88,8 @@ public class ClinicalAndTestingCohortsAction extends BaseAction implements Sessi
 	
 	private String discoveryPhene;
 	private String discoveryPheneTable;
-	private Integer discoveryLowCutoff;
-	private Integer discoveryHighCutoff;
+	private Double discoveryLowCutoff;
+	private Double discoveryHighCutoff;
 	
 	private String percentInValidationCohort;
 	
@@ -864,19 +864,19 @@ public class ClinicalAndTestingCohortsAction extends BaseAction implements Sessi
         this.discoveryPheneTable = discoveryPheneTable;
     }
 
-    public Integer getDiscoveryLowCutoff() {
+    public Double getDiscoveryLowCutoff() {
         return discoveryLowCutoff;
     }
 
-    public void setDiscoveryLowCutoff(Integer discoveryLowCutoff) {
+    public void setDiscoveryLowCutoff(Double discoveryLowCutoff) {
         this.discoveryLowCutoff = discoveryLowCutoff;
     }
 
-    public Integer getDiscoveryHighCutoff() {
+    public Double getDiscoveryHighCutoff() {
         return discoveryHighCutoff;
     }
 
-    public void setDiscoveryHighCutoff(Integer discoveryHighCutoff) {
+    public void setDiscoveryHighCutoff(Double discoveryHighCutoff) {
         this.discoveryHighCutoff = discoveryHighCutoff;
     }
